move snapshot tests to correct directory
